% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureScaling.R
\name{FeatureScaling}
\alias{FeatureScaling}
\title{Transform a numeric vector into a normalised vector of values between 0 and 1}
\usage{
FeatureScaling(x)
}
\arguments{
\item{Vector}{A Vector of numeric values.}
}
\value{
A normalised vector of numeric values between 0 and 1.
}
\description{
This function will return a vector between 0 and 1.
}
\examples{
FeatureScaling(c(252,325,39,103))
}
