% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t11.R
\name{t11}
\alias{t11}
\title{Transform categorical to normalised numeric on the Iowa houses db Neighborhoods}
\usage{
t11(x)
}
\arguments{
\item{neighborhood}{A valid Iowa homes neighborhood.}
}
\value{
0, 0.5 or 1
}
\description{
This function will return a number between 0 and 1.
This depends on whether the neighborhood is high class, medium class, or lower class.
}
\examples{
t11("Northridge")
}
