% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t21.R
\name{t21}
\alias{t21}
\title{Transform categorical to normalised numeric on a levelled rating system.}
\usage{
t21(x)
}
\arguments{
\item{Rating}{A valid rating, either "Excellent", "Good", "Average/Typical", "Fair" or "Poor"}
}
\value{
0, 0.25, 0.5, 0.75, 1
}
\description{
This function will return a number between 0 and 1.
This depends on whether the entry is Excellent(1) to Poor(0).
}
\examples{
t21("Excellent")
}
